Update install & provisioning scripts

+++ scripts/void/userifc.sh
@@ -1,50 +1,56 @@
 #!/bin/sh -x
 
 ## scripts/userifc.sh
+export CHOICE_DESKTOP=${1:-xfce}
 set +e
 
-export CHOICE_DESKTOP=${1:-xfce}
+snapshot_name=pre_userifc-$(date -u "+%Y%m%d") \
+  sh $(dirname ${0})/upgradepkgs.sh snapshot
 
-xbps-install -S ; xbps-install -uy xbps ; xbps-install -uy
-. /root/init/void/distro_pkgs.ini
-case ${CHOICE_DESKTOP} in
-	lxqt) pkgs_var="${pkgs_displaysvr_xorg} ${pkgs_deskenv_lxqt}" ;;
-	*) pkgs_var="${pkgs_displaysvr_xorg} ${pkgs_deskenv_xfce}" ;;
-esac
+xbps-install -S ; xbps-install -uy xbps
+#xbps-install -Duy ; xbps-install -uy
+. /root/scripts/distro_pkgs.ini
+echo ${pkgs_displaysvr} ${pkgs_deskenv_${CHOICE_DESKTOP}}
 
-for pkgX in ${pkgs_var} ; do
-	xbps-install -y -D ${pkgX} ;
+read -p "Enter 'y' to continue [nY]: " response
+#if [ "n" = "${response}" ] || [ "N" = "${response}" ] ; then
+if [ -n "$(echo ${response} | grep -e '^[Nn].*')" ] ; then
+  exit ;
+fi
+#xbps-install [-D] -y pkg0 .. pkgN # ERR, doesn't skip missing
+for pkgX in ${pkgs_displaysvr} ${pkgs_deskenv_${CHOICE_DESKTOP}} ; do
+  xbps-install -y ${pkgX} ;
 done
-for pkgX in ${pkgs_var} ; do
-	xbps-install -y ${pkgX} ;
+
+
+if [ -z "$(grep -e 'dbus-uuidgen --ensure' /etc/rc.local)" ] ; then
+  #echo /usr/bin/dbus-uuidgen --ensure=/etc/machine-id >> /etc/rc.local ;
+  echo /usr/bin/dbus-uuidgen --ensure >> /etc/rc.local ;
+  chmod +x /etc/rc.local ;
+fi
+mkdir -p /etc/X11/xorg.conf.d
+for confX in 10-evdev.conf 40-libinput.conf ; do
+  cp -an /usr/share/X11/xorg.conf.d/${confX} /etc/X11/xorg.conf.d/ ;
 done
-sleep 3
-
-case ${CHOICE_DESKTOP} in
-	lxqt) ln -s /etc/sv/sddm /etc/runit/runsvdir/default/ ;;
-	*) #mv /etc/lightdm /etc/lightdm.old ;
-	  ln -s /etc/sv/lightdm /etc/runit/runsvdir/default/ ;;
-esac
-ln -s /etc/sv/dbus /etc/runit/runsvdir/default/
-ln -s /etc/sv/polkitd /etc/runit/runsvdir/default/
-chmod 1777 /tmp
-
-# enable touchpad tapping
-sed -i '/MatchIsTouchpad/a \ \ \ \ \ \ \ \ Option "Tapping" "on"' \
-  /usr/share/X11/xorg.conf.d/10-evdev.conf
-sed -i '/MatchIsTouchpad/a \ \ \ \ \ \ \ \ Option "Tapping" "on"' \
-  /usr/share/X11/xorg.conf.d/40-libinput.conf
-## ??? egrep -i 'synap|alps|etps|elan' /proc/bus/input/devices
-#libinput list-devices ; xinput --list
-#xinput list-props XX [; xinput disable YY] # by id, list-props or disable
-#xinput set-prop <deviceid|devicename> <deviceproperty> <value>
-
-# update XDG user dir config
-export LANG=en_US.UTF-8 ; export CHARSET=UTF-8
-echo 'BIN=bin' >> /etc/xdg/user-dirs.defaults
-xdg-user-dirs-update ; sleep 3
-
-sed -i 's|nomodeset | |' /etc/default/grub
-sed -i 's|text | |' /etc/default/grub
+sed -i -e 's|nomodeset ||g' -e 's|text ||g' -e 's|xdriver=vesa ||g' /etc/default/grub
 grub-mkconfig -o /boot/grub/grub.cfg
 
+
+sh /root/init/common/misc_config.sh cfg_xdguserdirs /etc/xdg
+sh /root/init/common/misc_config.sh cfg_xorgtouchpad /etc/X11/xorg.conf.d
+
+set +e ; set +u
+echo "Enable|disable services" ; sleep 3
+for svc in ${uiservices_enabled_${CHOICE_DESKTOP}} ; do
+  ln -s /etc/sv/${svc} /etc/runit/runsvdir/default/ || true ;
+  #ln -s /etc/sv/${svc} /var/service/ || true ;
+done
+for svc in ${uiservices_disabled_${CHOICE_DESKTOP}} ; do
+  rm /etc/runit/runsvdir/default/${svc} || true ;
+  #rm /var/service/${svc} || true ;
+done
+
+
+set +e
+## scripts/cleanup.sh
+xbps-remove -oOy
